Fix warnings, based on a patch from Marc Haisenko.

+++ PSMTabBarControl/source/NSBezierPath_AMShading.m
@@ -11,7 +11,7 @@
 
 @implementation NSBezierPath (AMShading)
 
-static void linearShadedColor(void *info, const float *in, float *out)
+static void linearShadedColor(void *info, const CGFloat *in, CGFloat *out)
 {
 	float *colors = info;
 	*out++ = colors[0] + *in * colors[8];
@@ -20,7 +20,7 @@ static void linearShadedColor(void *info, const float *in, float *out)
 	*out++ = colors[3] + *in * colors[11];
 }
 
-static void bilinearShadedColor(void *info, const float *in, float *out)
+static void bilinearShadedColor(void *info, const CGFloat *in, CGFloat *out)
 {
 	float *colors = info;
 	float factor = (*in)*2.0;
@@ -141,8 +141,8 @@ static void bilinearShadedColor(void *info, const float *in, float *out)
 	// draw gradient
 	colorspace = CGColorSpaceCreateDeviceRGB();
 	size_t components = 1 + CGColorSpaceGetNumberOfComponents(colorspace);
-	static const float  domain[2] = {0.0, 1.0};
-	static const float  range[10] = {0, 1, 0, 1, 0, 1, 0, 1, 0, 1};
+	static const CGFloat  domain[2] = {0.0, 1.0};
+	static const CGFloat  range[10] = {0, 1, 0, 1, 0, 1, 0, 1, 0, 1};
 	//static const CGFunctionCallbacks callbacks = {0, &bilinearShadedColor, NULL};
 	
 	// Create a CGFunctionRef that describes a function taking 1 input and kChannelsPerColor outputs.
